requirements.txt
torch==1.13.1
torchaudio==0.13.1
auraloss==0.4.0
cython==3.0.10
einops
debugpy==1.8.0
fairseq==0.12.2
filelock~=3.12.2
h5py
kaldiio~=2.18.0
librosa==0.8.1
matplotlib~=3.4.3
nltk==3.8.1
numpy
pathlib~=1.0.1
pyyaml~=6.0
scikit-learn
scipy~=1.7.1
setuptools==65.6.3
six==1.16.0
soundfile~=0.10.3.post1
sox
tensorboard
tensorboardx~=2.5.1
tqdm~=4.62.3
transformers==4.42.3
